docker-compose/api/requirements.txt
#pip==21.1.3
requests==2.25.1
Flask==1.1.2
Flask-RESTful==0.3.8
jsonschema==3.2.0
mysql-connector-python==8.0.24
pandas
scikit-learn